+++ integrations/resend/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @trigger.dev/resend
 
+## 2.1.9
+
+### Patch Changes
+
+- Updated dependencies [9a187f9e]
+- Updated dependencies [2e9452ab]
+  - @trigger.dev/sdk@2.1.9
+  - @trigger.dev/integration-kit@2.1.9
+
+## 2.1.8
+
+### Patch Changes
+
+- 6a992a19: First release of `@trigger.dev/replicate` integration with remote callback support.
+- Updated dependencies [6a992a19]
+- Updated dependencies [ab9e4a98]
+- Updated dependencies [ab9e4a98]
+  - @trigger.dev/sdk@2.1.8
+  - @trigger.dev/integration-kit@2.1.8
+
 ## 2.1.7
 
 ### Patch Changes